Created a scene and finished character movement
I followed the game design document and built a simple scene. I added a yellow cube as the counter and a red cube to represent an ingredient box from which players can take ingredients. One thing to note is that when the player's speed exceeds around 11, the collider does not work properly, causing the character to run through objects. For now, I am using a capsule, a cube, and a sphere to represent the characters.