Add outline and interactions on objects
When a player character "looks" at an object, the object is outlined and
can be used. We do this with a raycast from the player character's head.
This doesn't address objects on top of other objects; ideally we want to
aim at the object that a plate is on and our interaction should be with
the plate and not with the object below.