Add chair models
I had to put the chair under an empty game object following the advice
from [1] since the pivot of the models are not at the center, they're
at the corner.
This makes working with the models a headache. I find out what the
actual center is with
`Debug.Log(gameObject.GetComponent<Renderer>().bounds.center)` and
offset the model under the empty game object with those values. It's
also much harder to work with the scripts since the collider is on the
model not on the empty game object, so the functions in Interactor hit
the model with the raycast, not the empty object. So you have to put the
scripts on the model. But other scripts like NPCSpawner rely on the
script being given to the empty object, so we have to give the Chair
script to both the empty object and its child.
It's just such a hassle; we should probably switch to a different asset
pack (the pivot is also wrong for the cabinet and stove).

[1] https://discussions.unity.com/t/cannot-rotate-an-object-around-the-center/100481