Add tomato soup recipe
Architecture for placing items in the pot is based on how the Code
Monkey video [1] handles placing items on plates.

Our PotUsableObject.TryAddIngredient() is different because we want to
open the possibility for multiple recipes and recipes that use an
ingredient more than once (e.g., our tomato soup recipe from Overcooked
which uses 3 chopped tomatoes).

Had to move the Player prefab's UsableObjectHoldPoint further out since
the collider of the pot was coming in contact with the player which
slowed movement.

[1] https://youtu.be/AmGSEH7QcDg&t=22088