Fix null exception when moving from cabinet to pot
When player is holding a pot and picking up ingredients from a cabinet,
the earlier if statement would have destroyed the object. The next if
statement then runs and fails because it tries to call a method on a
destroyed object.
Fix it by using `else if` so we don't call the second if statement if
the first condition passed.