Untag Chair prefab and use prefab defaults in scene
Don't use tags to identify objects; it's very fragile. We should also
have all the information needed in the Glass Table prefab, rather than
leaving the fields of the Glass Table script empty and then filling them
once we add a Glass Table to the scene.

Don't know what's te purpose of the Canvas field in GlassTable, so
commented out for now.