Fixed some npc-related errors
This commit should include all the fixes for the NPC code. I will not make any new updates to this branch unless other bugs and errors are found. If you think this code is okay to merge, feel free to merge it into the main branch. Then we can work on other things.