Move chairs around Glass Table
When the chairs and table are too close to each other, the collider of
the player overlaps with the other colliders incorrectly which is why
the outline of the glass table would sometimes disable when walking in
between the table and chairs.
Moving the chairs further out provides more space and prevents the
colliders from overlapping and breaking.