Make NPC collider smaller and move spawn location
The NPC collider being large and the spawn location being too close to
the wall is what was pushing the players into the wall when they moved
from east to west along the wall.
Make the NPC collider smaller for more intuitive collisions and move the
spawn location even further out to prevent this from happening.