Lower refrigerator collider
We activate a collider in front of the player above where they hold
objects when they are holding an object to prevent them from sticking
objects through walls. But this means when the player picks up an object
from the fridge, that collider will collide with the fridge's and push
the player away.
Work around this by lowering the fridge's collider. There shouldn't be
any objects that will make it that high anyway.