Don't highlight particle system when cooked/burned
Putting the Outline script on the top object will mesh all children
together. We only want the pot/pan to and its stuff to be outlined, not
the particle system, so we need to move the Outline script to an object
that contains the objects to be highlighted as children while having the
particle system outside of that object.