Reset currentLine in DialogManager at showing
If you show a dialog and then show another before calling CloseDialog(),
the currentLine would not be 0 when the next dialog starts. The UI shows
line 0 of the new dialog, but trying to advance could cancel the dialog
before the dialog finishes because the currentLine is ahead of the total
line number.